doparti-backend/requirements.txt
fastapi
uvicorn
requests
python-dateutil
pytz